requirements.txt
# Зависимости для HR Assistant Telegram Bot

# Telegram Bot API
python-telegram-bot==21.0.1

# Anthropic Claude API
anthropic==0.39.0

# Переменные окружения
python-dotenv==1.0.0
